Split dartpy simulation bindings into parallel-compilable TUs
simulation/module.cpp hosted all ~624 nanobind bindings (~3983 lines) in a
single void defSimulationModule(nb::module_&) translation unit. In Debug
(-O0 -g) that one TU compiled for 2h41m+ and timed out the required CI
"Debug Tests" job; because it is a single TU, -j parallelism could not help.

Split the bindings -- mechanically and without reordering -- across six
module_*.cpp parts (core, frames, bodies, compute, deformable_io, world) that
compile in parallel. module.cpp is now a thin coordinator that calls
defSimPart*(m) in the original registration order, which is load-bearing
because nanobind registers types in execution order and later parts reference
types registered by earlier parts. Shared file-local helpers (the validate*/
to*/make* functions and the Py* policy wrappers) moved into a new internal
header simulation/module_detail.hpp as inline definitions so every part shares
one ODR-safe copy.

sim::RigidBody is registered in the frames part (so earlier Joint bindings can
reference it) and its methods are attached in the bodies part via
nb::borrow<nb::class_<...>>(nb::type<sim::RigidBody>()); re-running
nb::class_<...>(m, "RigidBody") would re-register the type and emit a
RuntimeWarning.

Pure mechanical move: the concatenation of the six part bodies equals the
original function body, and nb::class_/nb::enum_/.def* registration counts are
unchanged. Validated with a Release dartpy build (-Werror clean), a clean
import (RuntimeWarning-as-error), and the simulation unit tests (134 passed,
8 pre-existing skips).
